feat(DTFS2-7392): extract supertest api types
