+++ trade-finance-manager-api/api-tests/types/file.ts
@@ -1,0 +1,4 @@
+export type File = {
+  fieldname: string;
+  filepath: string;
+};
